Playground.java: fixed matrix conversion (firestore returns Long not Integer)

+++ code/android-native/app/src/main/java/com/argame/model/data_structures/tic_tac_toe_game/ITicTacToeGame.java
@@ -32,7 +32,7 @@ public interface ITicTacToeGame {
 
     boolean isMyTurn();
 
-    int getUserPiece();
+    long getUserPiece();
 
-    List<Integer> getMatrix();
+    List<Long> getMatrix();
 }
